move queue and result set into SEA specific dir
